requirements.txt
MergePythonClient
langchain
openai
python-dotenv
huggingface_hub
streamlit
gradio
evals
ipykernel
pyautogen
instructor
